BACKEND/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
Django==5.0.4
django-allauth==0.63.3
django-cors-headers==4.3.1
django-ninja==1.1.0
idna==3.7
oauthlib==3.2.2
psycopg==3.1.19
psycopg-binary==3.1.19
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
PyJWT==2.8.0
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.1
